[Feature/FP-19] 데이터 베이스(Mysql)모듈 분리
