+++ finpik-infra/build.gradle
@@ -2,32 +2,18 @@ jar.enabled = true
 
 dependencies {
     implementation project(":finpik-domain")
-
-    //jpa
-    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-
-    //mysql
-    runtimeOnly 'com.mysql:mysql-connector-j'
+    implementation project(":finpik-infra:mysql")
 
     //h2
     runtimeOnly 'com.h2database:h2'
 
-    //querydsl
-    implementation("com.querydsl:querydsl-jpa:5.0.0:jakarta")
-    annotationProcessor("com.querydsl:querydsl-apt:5.0.0:jakarta")
-    annotationProcessor("jakarta.persistence:jakarta.persistence-api:3.1.0")
-    annotationProcessor("jakarta.annotation:jakarta.annotation-api:2.1.1")
-
     //kafka
-    implementation("org.springframework.kafka:spring-kafka")
+    implementation 'org.springframework.kafka:spring-kafka'
     testImplementation 'org.springframework.kafka:spring-kafka-test'
 
     //redis
     implementation 'org.springframework.boot:spring-boot-starter-data-redis'
     implementation 'org.springframework.boot:spring-boot-starter-cache'
-
-    //kafka-kraft
-    implementation 'org.springframework.kafka:spring-kafka'
 
     //jwt
     implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
