+++ finpik-infra/mysql/build.gradle
@@ -1,0 +1,18 @@
+jar.enabled = true
+
+dependencies {
+    implementation project(":finpik-domain")
+
+    //jpa
+    api 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    //mysql
+    runtimeOnly 'com.mysql:mysql-connector-j'
+
+    //querydsl
+    implementation("com.querydsl:querydsl-jpa:5.0.0:jakarta")
+    annotationProcessor("com.querydsl:querydsl-apt:5.0.0:jakarta")
+    implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
+    annotationProcessor("jakarta.persistence:jakarta.persistence-api:3.1.0")
+    annotationProcessor("jakarta.annotation:jakarta.annotation-api:2.1.1")
+}
