+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/history/userproductview/UserProductViewHistoryJpaRepository.java
@@ -1,0 +1,11 @@
+package finpik.jpa.repository.history.userproductview;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import finpik.entity.history.userproductview.UserProductViewHistoryEntity;
+
+public interface UserProductViewHistoryJpaRepository
+        extends
+            JpaRepository<UserProductViewHistoryEntity, Long>,
+            CustomUserProductViewHistoryRepository {
+}
