+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/history/userproductview/projection/RelatedLoanProductProjection.java
@@ -1,0 +1,10 @@
+package finpik.jpa.repository.history.userproductview.projection;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+public class RelatedLoanProductProjection {
+    private Long productId;
+}
