+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/loanproduct/CustomLoanProductRepository.java
@@ -1,0 +1,9 @@
+package finpik.jpa.repository.loanproduct;
+
+import java.util.Optional;
+
+import finpik.entity.loanproduct.LoanProductEntity;
+
+public interface CustomLoanProductRepository {
+    Optional<LoanProductEntity> findByIdWithDescription(Long loanProductId);
+}
