+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/loanproduct/LoanProductJpaRepository.java
@@ -1,0 +1,8 @@
+package finpik.jpa.repository.loanproduct;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import finpik.entity.loanproduct.LoanProductEntity;
+
+public interface LoanProductJpaRepository extends JpaRepository<LoanProductEntity, Long>, CustomLoanProductRepository {
+}
