+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/profile/ProfileEntityJpaRepository.java
@@ -1,0 +1,12 @@
+package finpik.jpa.repository.profile;
+
+import java.util.List;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import finpik.entity.profile.ProfileEntity;
+import finpik.entity.user.UserEntity;
+
+public interface ProfileEntityJpaRepository extends JpaRepository<ProfileEntity, Long> {
+    List<ProfileEntity> findByUser(UserEntity userEntity);
+}
