+++ finpik-infra/mysql/src/main/java/finpik/jpa/repository/user/UserEntityJpaRepository.java
@@ -1,0 +1,11 @@
+package finpik.jpa.repository.user;
+
+import java.util.Optional;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import finpik.entity.user.UserEntity;
+
+public interface UserEntityJpaRepository extends JpaRepository<UserEntity, Long> {
+    Optional<UserEntity> findByEmail(String email);
+}
