--- finpik-infra/src/main/java/finpik/db/jpa/repository/history/userproductview/projection/RelatedLoanProductProjection.java
@@ -1,10 +1,0 @@
-package finpik.db.jpa.repository.history.userproductview.projection;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-public class RelatedLoanProductProjection {
-    private Long productId;
-}
