--- finpik-infra/src/main/java/finpik/db/jpa/repository/loanproduct/CustomLoanProductRepository.java
@@ -1,9 +1,0 @@
-package finpik.db.jpa.repository.loanproduct;
-
-import java.util.Optional;
-
-import finpik.db.entity.loanproduct.LoanProductEntity;
-
-public interface CustomLoanProductRepository {
-    Optional<LoanProductEntity> findByIdWithDescription(Long loanProductId);
-}
