--- finpik-infra/src/main/java/finpik/db/jpa/repository/loanproduct/LoanProductJpaRepository.java
@@ -1,8 +1,0 @@
-package finpik.db.jpa.repository.loanproduct;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import finpik.db.entity.loanproduct.LoanProductEntity;
-
-public interface LoanProductJpaRepository extends JpaRepository<LoanProductEntity, Long>, CustomLoanProductRepository {
-}
