+++ settings.gradle
@@ -6,3 +6,4 @@ include 'finpik-domain'
 include 'finpik-external'
 
 include 'finpik-infra'
+include 'finpik-infra:mysql'
